requirements.txt
aiohappyeyeballs   2.4.4
aiohttp            3.11.11
aiosignal          1.3.2
asyncio            3.4.3
attrs              25.1.0
certifi            2024.12.14
cffi               1.17.1
charset-normalizer 3.4.1
dnspython          2.7.0
frozenlist         1.5.0
h11                0.14.0
idna               3.10
motor              3.7.0
multidict          6.1.0
outcome            1.3.0.post0
pip                25.0
propcache          0.2.1
pycparser          2.22
pymongo            4.11
PySocks            1.7.1
python-dotenv      1.0.1
redis              5.2.1
requests           2.32.3
selenium           4.28.1
sniffio            1.3.1
sortedcontainers   2.4.0
trio               0.28.0
trio-websocket     0.11.1
typing_extensions  4.12.2
urllib3            2.3.0
websocket-client   1.8.0
wsproto            1.2.0
yarl               1.18.3